+++ src/main/java/duke/storage/DukeCompileTimeException.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.storage;
 
 public class DukeCompileTimeException extends Exception {
     public DukeCompileTimeException(String errorMessage) {
