+++ src/main/java/duke/storage/DukeIOException.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.storage;
 
 public class DukeIOException extends DukeCompileTimeException {
     public DukeIOException(String errorMessage) {
